pyproject.toml
[tool.poetry]
name = "naskpy"
version = "0.1.0"
description = "Tools, modules and functions that I use regularly while coding in Python"
authors = ["Mehdi Nassim KHODJA <khodjamehdinassim@gmail.com>"]
maintainers = ["Mehdi Nassim KHODJA <khodjamehdinassim@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/naskio/naskpy"
repository = "https://github.com/naskio/naskpy"
documentation = "https://py.nask.io/"
keywords = ["Python", "Python3", "library","utilities"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]
[tool.poetry.urls]
issues = "https://github.com/naskio/naskpy/issues"


[tool.poetry.dependencies]
python = ">=3.7,<3.10"
pendulum = "^2.1.2"
Pillow = "8.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-asyncio = "^0.14.0"
# coverage = "^5.5"
# tox = "^3.23.0"
black = "^20.8b1"
flake8 = "^3.9.2" # Your Tool For Style Guide Enforcement
mypy = "^0.910"
pre-commit = "^2.15.0"
isort = "^5.9.3"
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"
pytest-randomly = "^3.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Pre-commit hooks related config.
[tool.pycln] # A formatter for finding and removing unused import statements
all = true

[mypy] # Optional Static Typing for Python
ignore_missing_imports = true
no_strict_optional = true

[tool.isort] # library to sort imports
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black] # The uncompromising Python code formatter
line-length = 88
